# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Desafio Coodesh",
        "description": "API Desafio Coodesh",
        "contact": {
            "email": "tss.labsi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/products/{code}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show product",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product data",
                "operationId": "updateProducts",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "code",
                                    "status",
                                    "imported_t",
                                    "url",
                                    "creator",
                                    "created_t",
                                    "last_modified_t",
                                    "product_name",
                                    "quantity",
                                    "brands",
                                    "categories",
                                    "labels",
                                    "cities",
                                    "purchase_places",
                                    "stores",
                                    "ingredients_text",
                                    "traces",
                                    "serving_size",
                                    "serving_quantity",
                                    "nutriscore_score",
                                    "nutriscore_grade",
                                    "main_category",
                                    "image_url"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "imported_t": {
                                        "type": "dateTime"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "creator": {
                                        "type": "string"
                                    },
                                    "created_t": {
                                        "type": "string"
                                    },
                                    "last_modified_t": {
                                        "type": "string"
                                    },
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    },
                                    "brands": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "string"
                                    },
                                    "labels": {
                                        "type": "string"
                                    },
                                    "cities": {
                                        "type": "string"
                                    },
                                    "purchase_places": {
                                        "type": "string"
                                    },
                                    "stores": {
                                        "type": "string"
                                    },
                                    "ingredients_text": {
                                        "type": "string"
                                    },
                                    "traces": {
                                        "type": "string"
                                    },
                                    "serving_size": {
                                        "type": "string"
                                    },
                                    "serving_quantity": {
                                        "type": "string"
                                    },
                                    "nutriscore_score": {
                                        "type": "string"
                                    },
                                    "nutriscore_grade": {
                                        "type": "string"
                                    },
                                    "main_category": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}